requirements.txt
absl-py==0.10.0
aiohttp==3.6.2
asn1crypto==0.24.0
astunparse==1.6.3
async-timeout==3.0.1
attrs==20.2.0
blis==0.4.1
cachetools==4.1.1
catalogue==1.0.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography==2.1.4
cycler==0.10.0
cymem==2.0.3
Cython==0.29.21
dataclasses==0.7
dill==0.3.2
dm-tree==0.1.5
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz
future==0.18.2
gast==0.3.3
gin-config==0.3.0
google-api-core==1.23.0
google-api-python-client==1.12.5
google-auth==1.22.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
google-cloud-bigquery==2.2.0
google-cloud-core==1.4.3
google-crc32c==1.0.0
google-pasta==0.2.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
grpcio==1.32.0
h5py==2.10.0
httplib2==0.18.1
idna==2.6
idna-ssl==1.1.0
importlib-metadata==2.0.0
importlib-resources==3.1.1
joblib==0.17.0
kaggle==1.5.9
Keras-Preprocessing==1.1.2
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.2.0
Markdown==3.2.2
matplotlib==3.3.2
multidict==4.7.6
murmurhash==1.0.2
numpy==1.18.5
oauthlib==3.1.0
opencv-python-headless==4.4.0.44
opt-einsum==3.3.0
pandas==1.1.3
Pillow==8.0.1
plac==1.1.3
preshed==3.0.2
promise==2.3
proto-plus==1.11.0
protobuf==3.13.0
psutil==5.7.3
py-cpuinfo==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycrypto==2.6.1
pygobject==3.26.1
pyparsing==2.4.7
python-apt==1.6.5+ubuntu0.3
python-dateutil==2.8.1
python-slugify==4.0.1
pytz==2020.1
pyxdg==0.25
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-learn==0.23.2
scipy==1.5.3
SecretStorage==2.3.1
sentencepiece==0.1.94
six==1.15.0
sklearn==0.0
slugify==0.0.1
spacy==2.3.2
srsly==1.0.2
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-addons==0.11.2
tensorflow-datasets==4.0.1
tensorflow-estimator==2.3.0
tensorflow-gpu==2.3.1
tensorflow-hub==0.9.0
tensorflow-metadata==0.24.0
tensorflow-model-optimization==0.5.0
termcolor==1.1.0
text-unidecode==1.3
tf-models-official==2.3.0
tf-slim==1.1.0
thinc==7.4.1
threadpoolctl==2.1.0
tqdm==4.51.0
typeguard==2.10.0
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.25.10
wasabi==0.8.0
Werkzeug==1.0.1
wrapt==1.12.1
yarl==1.6.0
zipp==3.2.0
